refactor(chat): extract Store<T> primitive to store/
