+++ apps/mesh/src/web/components/chat/hooks/thread-connection.ts
@@ -42,7 +42,10 @@ import {
 } from "ai";
 import type { ToolApprovalLevel } from "@/web/hooks/use-preferences";
 import type { SimpleModeTier } from "@/tools/organization/schema";
+import { Store } from "../store/store-primitive";
 import type { ChatMode } from "../types";
+
+export { Store };
 
 // ─── Request options (wire payload alongside `messages`) ─────────────────────
 
@@ -85,30 +88,8 @@ export type SubmitAction =
     };
 
 // ─── Hand-rolled emitter ─────────────────────────────────────────────────────
-
-/**
- * Per-update emitter (does not batch across SSE chunks). React subscribes via
- * useSyncExternalStore.
- */
-export class Store<T> {
-  private subs = new Set<() => void>();
-  constructor(private value: T) {}
-  get = (): T => this.value;
-  set = (next: T): void => {
-    if (Object.is(next, this.value)) return;
-    this.value = next;
-    this.subs.forEach((s) => s());
-  };
-  update = (fn: (prev: T) => T): void => {
-    this.set(fn(this.value));
-  };
-  subscribe = (s: () => void): (() => void) => {
-    this.subs.add(s);
-    return () => {
-      this.subs.delete(s);
-    };
-  };
-}
+// `Store<T>` lives in ../store/store-primitive; re-exported above for callers
+// that import it from this module.
 
 // ─── Observer ────────────────────────────────────────────────────────────────
